Moves Layout component to shared
