+++ src/components/shared/Layout/Layout.js
@@ -1,10 +1,9 @@
 import React from 'react';
 import Helmet from 'react-helmet';
 import cx from 'classnames';
-import Header from '../Header';
-
+import Header from '../../Header';
+import LastLessonProvider from '../../../providers/LastLessonProvider';
 import './Layout.scss';
-import LastLessonProvider from '../../providers/LastLessonProvider';
 
 const Layout = ({children, title, description, isFullBleed}) => {
   const layoutClass = cx('Layout', {
