+++ src/templates/categories-list-template.js
@@ -1,11 +1,11 @@
-import React from "react";
-import { Link, graphql } from "gatsby";
-import kebabCase from "lodash/kebabCase";
-import { Container, Row, Col } from "react-grid-system";
-import PropTypes from "prop-types";
+import React from 'react';
+import { Link, graphql } from 'gatsby';
+import kebabCase from 'lodash/kebabCase';
+import { Container, Row, Col } from 'react-grid-system';
+import PropTypes from 'prop-types';
 import { SITE_METADATA_PROP_TYPE } from '../constants/propTypes';
-import Layout from "../components/Layout";
-import Page from "../components/Page";
+import Layout from '../components/shared/Layout';
+import Page from '../components/Page';
 
 const CategoriesListTemplate = ({ data }) => {
   const { title, subtitle } = data.site.siteMetadata;
