+++ src/templates/course-template.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { graphql } from 'gatsby';
-import Layout from '../components/Layout';
+import Layout from '../components/shared/Layout';
 import Course from '../components/Course';
 import Page from '../components/Page';
