+++ src/templates/not-found-template.js
@@ -1,8 +1,8 @@
-import React from "react";
-import { Container, Row, Col } from "react-grid-system";
-import { graphql } from "gatsby";
-import Layout from "../components/Layout";
-import Page from "../components/Page";
+import React from 'react';
+import { Container, Row, Col } from 'react-grid-system';
+import { graphql } from 'gatsby';
+import Layout from '../components/shared/Layout';
+import Page from '../components/Page';
 
 const NotFoundTemplate = ({ data }) => {
   const { title, subtitle } = data.site.siteMetadata;
